+++ draft-js-mention-plugin/src/MentionSuggestions/index.js
@@ -223,6 +223,10 @@ export default class MentionSuggestions extends Component {
     this.setState({
       isActive: true,
     });
+
+    if (this.props.onOpen) {
+      this.props.onOpen();
+    }
   };
 
   closeDropdown = () => {
@@ -239,6 +243,10 @@ export default class MentionSuggestions extends Component {
     this.setState({
       isActive: false,
     });
+
+    if (this.props.onClose) {
+      this.props.onClose();
+    }
   };
 
   render() {